Crate icu_plurals_data

Data for the icu_plurals crate

This data was generated with CLDR version 45.0.0, ICU version icu4x/2024-05-16/75.x, and LSTM segmenter version v0.1.0.

Macros

impl_plurals_cardinal_v1

Implement DataProvider<CardinalV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_plurals_ordinal_v1

Implement DataProvider<OrdinalV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_plurals_ranges_v1

Implement DataProvider<PluralRangesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impliterable_plurals_cardinal_v1

Implement IterableDataProvider<CardinalV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_plurals_ordinal_v1

Implement IterableDataProvider<OrdinalV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_plurals_ranges_v1

Implement IterableDataProvider<PluralRangesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

make_provider

Marks a type as a data provider. You can then use macros like impl_core_helloworld_v1 to add implementations.